requirements.txt
Flask
wikipedia
requests
beautifulsoup4
googlesearch-python
fpdf
pandas
openai
